backend/pyproject.toml
[project]
name = "Auto_Bangumi"
version = "3.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiolimiter>=1.2.1",
  "alembic>=1.17.2",
  "bcrypt==4.2.0",
  "beautifulsoup4>=4.13.4",
  "bencode-py>=4.0.0",
  "dotenv>=0.9.9",
  "fastapi>=0.116.1",
  "httpx[http2,socks]>=0.28.1",
  "jinja2>=3.1.6",
  "packaging>=25.0",
  "passlib>=1.7.4",
  "pydantic-settings>=2.0.0",
  "python-jose>=3.5.0",
  "python-multipart>=0.0.20",
  "sqlmodel>=0.0.24",
  "sse-starlette>=3.0.2",
  "urllib3>=2.5.0",
  "uvicorn>=0.35.0",
]
[tool.ruff]
select = [
  # pycodestyle(E): https://beta.ruff.rs/docs/rules/#pycodestyle-e-w
  "E",
  # Pyflakes(F): https://beta.ruff.rs/docs/rules/#pyflakes-f
  "F",
  # isort(I): https://beta.ruff.rs/docs/rules/#isort-i
  "I",
]
ignore = [
  # E501: https://beta.ruff.rs/docs/rules/line-too-long/
  'E501',
  # F401: https://beta.ruff.rs/docs/rules/unused-import/
  # avoid unused imports lint in `__init__.py`
  'F401',
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
  "A",
  "B",
  "C",
  "D",
  "E",
  "F",
  "G",
  "I",
  "N",
  "Q",
  "S",
  "T",
  "W",
  "ANN",
  "ARG",
  "BLE",
  "COM",
  "DJ",
  "DTZ",
  "EM",
  "ERA",
  "EXE",
  "FBT",
  "ICN",
  "INP",
  "ISC",
  "NPY",
  "PD",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "TID",
  "TRY",
  "UP",
  "YTT",
]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]
per-file-ignores = {}

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.black]
line-length = 120
target-version = ['py311']

[dependency-groups]
dev = ["pytest>=8.4.1", "pytest-asyncio>=1.1.0"]
